+++ ds_tas/engine/__init__.py
@@ -1,0 +1,7 @@
+"""
+Core engine components go in this module.
+
+The end user shouldn't need to play around with these.
+"""
+
+from .tas_engine import tas
